[sys_profile] Fixed hitpoints not detected for some RHS vehicles #286
Some RHS vehicles were not returning any hitpoints for use when calculating virtualised damage. [(configfile >> "CfgVehicles" >> _vehicleClass >> "HitPoints"),0] call BIS_fnc_returnChildren; is a much slower method of getting the hitpoints but it seems to be more thorough and works on all RHS vehicles. This is only called when the original method doesn't work. 

This solves the "Undefined variable in expression: _randomhitpoint" error.