[X LIB] New feature: Disguise as the enemy
Disguises given unit as the given enemy target if there are no nearby hostile units watching. Mind also civilians could be hostile.
Cover will be blown if unit is near enemies that can also see the unit and the unit acts obvious (shooting, running, etc.) or if the uniform is changed.

- Will add an action "Disguise as the enemy" on the local player client if called via "call ALIVE_fnc_disguise" (f.e. in mission init.sqf).
- Action will persist on respawn.
- Runs locally on players client only and therefore won't affect server performance at all.

TBD: Implement as a player option