Added isnil check on attempts to retrieve _profile
This is an attempt to reduce the effects of the race condition.

Since units profiles are unregistered upon death by an event handler, opcom and the intelligence may try to act upon already unregistered profiles.

This would generate an error spam on fnc_G2.sqf and also rarely stall opcom.

Since the profiles are returned as new instances, the simple and needed workaround is to sanitize _profile, checking if the returned value was null before trying to get values from it.